backend/requirements.txt
# FastAPI and Server
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
httpx==0.27.2
sse-starlette==2.1.3
pydantic==2.9.2
pydantic-settings==2.5.2

# LangChain and LangGraph
langchain
langchain-openai
langchain-community
langgraph

# Hedera SDK (Official Hiero SDK)
hiero-sdk-python

# External APIs
tavily-python==0.5.0

# Utilities
aiohttp==3.10.10
python-json-logger==2.0.7

